moved LocalAssetAdapter to common-runtime
