+++ runtime/common/Cargo.toml
@@ -1,32 +1,34 @@
 [package]
-name = 'runtime-common'
-version = '0.0.1'
-description = 'A node of the peaq network.'
-authors = ['peaq network <https://github.com/peaqnetwork>']
-homepage = 'https://peaq.network/'
-edition = '2021'
+name = "runtime-common"
+version = "0.0.1"
+description = "A node of the peaq network."
+authors = ["peaq network <https://github.com/peaqnetwork>"]
+homepage = "https://peaq.network/"
+edition = "2021"
 publish = false
 
 
 [dependencies]
-sp-runtime = { workspace = true, default-features = false }
-sp-std = { workspace = true, default-features = false }
-pallet-transaction-payment = { workspace = true, default-features = false }
 frame-support = { workspace = true, default-features = false }
 frame-system = { workspace = true, default-features = false }
-
-peaq-primitives-xcm = { path = "../../primitives/xcm", default-features = false }
 orml-tokens = { workspace = true, default-features = false }
 orml-traits = { workspace = true, default-features = false }
+pallet-transaction-payment = { workspace = true, default-features = false }
+peaq-primitives-xcm = { path = "../../primitives/xcm", default-features = false }
+sp-runtime = { workspace = true, default-features = false }
+sp-std = { workspace = true, default-features = false }
+zenlink-protocol = { workspace = true, default-features = false }
+
 
 [features]
 default = ["std"]
 std = [
-	'sp-runtime/std',
-	'pallet-transaction-payment/std',
+	"sp-runtime/std",
+	"pallet-transaction-payment/std",
 	"peaq-primitives-xcm/std",
-	'frame-support/std',
+	"frame-support/std",
 	"orml-tokens/std",
 	"orml-traits/std",
-	'frame-system/std',
+	"frame-system/std",
+	"zenlink-protocol/std",
 ]
