+++ runtime/common/src/lib.rs
@@ -2,18 +2,26 @@
 #![recursion_limit = "256"]
 
 
-use sp_std::marker::PhantomData;
-use sp_runtime::Perbill;
-
 use frame_support::{
 	parameter_types,
+	pallet_prelude::{DispatchError, DispatchResult},
 	traits::Get,
 };
-use orml_traits::currency::MutationHooks;
+use orml_traits::{
+	MultiCurrency,
+	currency::MutationHooks,
+};
+use sp_std::marker::PhantomData;
+use sp_runtime::Perbill;
+use zenlink_protocol::{
+	AssetBalance, AssetId as ZenlinkAssetId, LocalAssetHandler,
+};
 
 
-/// Balance of an account.
-pub type Balance = peaq_primitives_xcm::Balance;
+// /// Balance of an account.
+// pub type Balance = peaq_primitives_xcm::Balance;
+
+use peaq_primitives_xcm::{Balance, CurrencyId};
 
 // Contracts price units.
 pub const TOKEN_DECIMALS: u32 = 18;
@@ -29,6 +37,7 @@ parameter_types! {
 }
 
 pub struct CurrencyHooks<T, DustAccount>(PhantomData<T>, DustAccount);
+
 impl<T, DustAccount> MutationHooks<T::AccountId, T::CurrencyId, T::Balance> for CurrencyHooks<T, DustAccount>
 where
 	T: orml_tokens::Config,
@@ -42,4 +51,98 @@ where
 	type PostTransfer = ();
 	type OnNewTokenAccount = ();
 	type OnKilledTokenAccount = ();
+}
+
+
+
+/// A local adaptor to convert between Zenlink-Assets and Peaq's local currency.
+pub struct LocalAssetAdaptor<Local>(PhantomData<Local>);
+
+impl<Local, AccountId> LocalAssetHandler<AccountId> for LocalAssetAdaptor<Local>
+where
+	Local: MultiCurrency<AccountId, CurrencyId = CurrencyId>,
+{
+	fn local_balance_of(asset_id: ZenlinkAssetId, who: &AccountId) -> AssetBalance {
+		if let Ok(currency_id) = asset_id.try_into() {
+			return TryInto::<AssetBalance>::try_into(Local::free_balance(currency_id, &who))
+				.unwrap_or_default();
+		}
+		AssetBalance::default()
+	}
+
+	fn local_total_supply(asset_id: ZenlinkAssetId) -> AssetBalance {
+		if let Ok(currency_id) = asset_id.try_into() {
+			return TryInto::<AssetBalance>::try_into(Local::total_issuance(currency_id))
+				.unwrap_or_default();
+		}
+		AssetBalance::default()
+	}
+
+	fn local_is_exists(asset_id: ZenlinkAssetId) -> bool {
+		let currency_id: Result<CurrencyId, ()> = asset_id.try_into();
+		match currency_id {
+			Ok(_) => true,
+			Err(_) => false,
+		}
+	}
+
+	fn local_transfer(
+		asset_id: ZenlinkAssetId,
+		origin: &AccountId,
+		target: &AccountId,
+		amount: AssetBalance,
+	) -> DispatchResult {
+		if let Ok(currency_id) = asset_id.try_into() {
+			Local::transfer(
+				currency_id,
+				&origin,
+				&target,
+				amount
+					.try_into()
+					.map_err(|_| DispatchError::Other("convert amount in local transfer"))?,
+			)
+		} else {
+			Err(DispatchError::Other("unknown asset in local transfer"))
+		}
+	}
+
+	fn local_deposit(
+		asset_id: ZenlinkAssetId,
+		origin: &AccountId,
+		amount: AssetBalance,
+	) -> Result<AssetBalance, DispatchError> {
+		if let Ok(currency_id) = asset_id.try_into() {
+			Local::deposit(
+				currency_id,
+				&origin,
+				amount
+					.try_into()
+					.map_err(|_| DispatchError::Other("convert amount in local deposit"))?,
+			)?;
+		} else {
+			return Err(DispatchError::Other("unknown asset in local transfer"));
+		}
+
+		Ok(amount)
+	}
+
+	fn local_withdraw(
+		asset_id: ZenlinkAssetId,
+		origin: &AccountId,
+		amount: AssetBalance,
+	) -> Result<AssetBalance, DispatchError> {
+		if let Ok(currency_id) = asset_id.try_into() {
+			Local::withdraw(
+				currency_id,
+				&origin,
+				amount
+					.try_into()
+					.map_err(|_| DispatchError::Other("convert amount in local withdraw"))?,
+			)?;
+		} else {
+			return Err(DispatchError::Other("unknown asset in local transfer"));
+		}
+
+		Ok(amount)
+	}
 }
